pyproject.toml
[project]
name = "pdf2ppdx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "cairosvg>=2.7.1",
    "pdf2image>=1.17.0",
    "python-pptx>=1.0.2",
]
